+++ src/openapiToTsJsonSchema.ts
@@ -4,7 +4,7 @@ import $RefParser from '@apidevtools/json-schema-ref-parser';
 import get from 'lodash.get';
 import {
   clearFolder,
-  makeJsonSchemaFiles,
+  makeTsJsonSchemaFiles,
   REF_SYMBOL,
   convertOpenApiToJsonSchema,
   convertOpenApiParameters,
@@ -117,9 +117,9 @@ export async function openapiToTsJsonSchema({
   const schemaMetaDataMap: SchemaMetaDataMap = new Map();
 
   /**
-   * Generate meta data for schemas which have been previously dereferenced.
-   * It happens for "import" and "keep" refHandling since they expect to find
-   * $ref schemas generated as standalone schemas
+   * Create meta data for $ref schemas which have been previously dereferenced.
+   * It happens only with "import" and "keep" refHandling since they expect
+   * $ref schemas to be generated no matter of
    */
   if (refHandling === 'import' || refHandling === 'keep') {
     for (const [ref, schema] of inlinedRefs) {
@@ -135,7 +135,9 @@ export async function openapiToTsJsonSchema({
     }
   }
 
-  // Generate schema meta info for user requested schemas
+  /**
+   * Create meta data for each output schema
+   */
   for (const definitionPath of definitionPathsToGenerateFrom) {
     const definitionSchemas = get(jsonSchema, definitionPath);
 
@@ -158,7 +160,7 @@ export async function openapiToTsJsonSchema({
     }
   }
 
-  await makeJsonSchemaFiles({
+  await makeTsJsonSchemaFiles({
     refHandling,
     schemaMetaDataMap,
   });
