+++ src/utils/makeTsJsonSchema/replacePlaceholdersWithRefs.ts
